Acorn ZIP files typically have extended information in them that preserves the Load/Exec address (amongst others).  When extracting these Zip files on a Unix machine that information is lost.
This program will parse a zip file and create a shell script (on stdout)
that will build ".inf" files that are compatible with the PiEconetBridge

We only extra the load/exec addresses from the zip file.
We hard code the owner of the file as SYST (user 0).  Directories are
given permission WR/ and files given permission LWR/R.

So the idea is you can unzip the Zip file and then run this program
to generate a script that would set the attributes

e.g
 cd /econet/0System
 mkdir ArthurLib
 cd ArthurLib
 unzip ..../NetLibA.zip
 ..../parse_acorn_zip.pl ..../NetLibA.zip | bash